Cleaned user route, exported some helper functions
