+++ routes/user.js
@@ -1,56 +1,23 @@
 var express = require('express');
 var userRoute = express.Router();
 
-var models = require('doorlock-models'); // Local link
+var models = require('doorlock-models');
 var User = models.User;
-
-function requestFilter(query) {
-
-  var filter = {}
-  var orderDirection = query.order || 'ASC';
-  var orderBy = query.orderBy || 'id';
-  var orderFirst = [];
-  var order = [];
-  orderFirst.push(orderBy,orderDirection);
-  order.push(orderFirst);
-  filter.order = order;
-  if (query.limit) { filter.limit = +query.limit }
-  if (query.offset) { filter.offset = +query.offset }
-
-  var where = {}
-  if (query.firstName) {
-    where = Object.assign(where, {
-      firstName: {
-        $like: '%' + query.firstName + '%'
-      }
-    });
-  }
-  if (query.lastName) {
-    where = Object.assign(where, {
-      lastName: {
-        $like: '%' + query.lastName + '%'
-      }
-    });
-  }
-
-  filter.where = where;
-
-  return filter;
-}
+var filterUser = require('../helpers/filterUser');
+var handleUserError = require('../helpers/handleUserError');
 
 userRoute.get('/', function(req,res,next) {
 
-  User.findAndCountAll(requestFilter(req.query))
-  .then( function(rows) {
-    res.json({ success: true, data: rows });
-  })
-  .catch( function(err) {
-    var getError = new Error('Internal server error');
-    next(getError);
-  });
+  User.findAndCountAll(filterUser(req.query))
+    .then( function(rows) {
+      res.json({ success: true, data: rows });
+    })
+    .catch( function(err) {
+      var getError = new Error('Internal server error');
+      next(getError);
+    });
 
 });
-
 
 userRoute.get('/findByID/:id', function(req,res,next) {
   if( !req.params.id ) {
@@ -62,48 +29,35 @@ userRoute.get('/findByID/:id', function(req,res,next) {
     User.findOne({
       where: { id: req.params.id }
     })
-    .then(function(user) {
-      res.json({
-        success: true,
-        data: user
+      .then(function(user) {
+        res.json({
+          success: true,
+          data: user
+        });
+      })
+      .catch( function(err) {
+        var getError = new Error('Internal server error');
+        next(getError);
       });
-    })
-    .catch( function(err) {
-      var getError = new Error('Internal server error');
-      next(getError);
-    });
   }
 });
 
 userRoute.post('/add', function(req, res, next) {
+
   User.create({
     firstName: req.body.firstName,
     lastName: req.body.lastName,
     userName: req.body.userName,
-	graduationYear: req.body.graduationYear,
+    graduationYear: req.body.graduationYear,
     privateEmail: req.body.privateEmail,
     mobile: req.body.mobile,
     studentCardId: req.body.studentCardId
   })
-  .then( function(row) {
-    res.json({success: true, data: row});
-  })
-  .catch( function(err) {
-    switch (err.name) {
-      case 'SequelizeUniqueConstraintError':
-        var UniqueConstraintError = new Error('Student card already in use');
-        UniqueConstraintError.status = 400;
-        next(UniqueConstraintError);
-        break;
-      case 'SequelizeValidationError':
-        var validationError = new Error('Validation error');
-        validationError.status = 400;
-        next(validationError);
-        break;
-      default:
-        next(new Error('Internal server error')); // return 500
-    }
-  });
+    .then( function(row) {
+      res.json({success: true, data: row});
+    })
+    .catch( function(err) { handleUserError(err,next) });
+
 });
 
 userRoute.put('/edit/:id', function(req, res, next) {
@@ -111,48 +65,37 @@ userRoute.put('/edit/:id', function(req, res, next) {
   User.update(req.body,
   {
     where: { id: req.params.id }
-  }).then( (affectedRows) => {
-    if(affectedRows) {
-      res.json({success: true, message: 'Member updated'})
-    }
-    else {
-      var notFound = new Error('Member not found');
-      notFound.status = 404;
-      next(notFound);
-    }
-  }).catch( (err) => {
-    switch (err.name) {
-      case 'SequelizeUniqueConstraintError':
-        var UniqueConstraintError = new Error('Student card already in use');
-        UniqueConstraintError.status = 400;
-        next(UniqueConstraintError);
-        break;
-      case 'SequelizeValidationError':
-        var validationError = new Error('Validation error');
-        validationError.status = 400;
-        next(validationError);
-        break;
-      default:
-        next(new Error('Internal server error')); // return 500
-    }
-  });
+  })
+    .then( (affectedRows) => {
+      if(affectedRows) {
+        res.json({success: true, message: 'Member updated'})
+      }
+      else {
+        var notFound = new Error('Member not found');
+        notFound.status = 404;
+        next(notFound);
+      }
+    })
+    .catch( (err) => { handleUserError(err,next) });
 
 });
 
 userRoute.delete('/delete/:id', function(req, res, next) {
 
-  User.destroy({where: { id: req.params.id } }).then( (affectedRows) => {
-    if(affectedRows) {
-      res.json({success: true, message: 'Member deleted'});
-    }
-    else {
-      var notFound = new Error('Member not found');
-      notFound.status = 404;
-      next(notFound);
-    }
-  }).catch( (err) => {
-    next(new Error('Internal server error')); // return 500
-  });
+  User.destroy({where: { id: req.params.id } })
+    .then( (affectedRows) => {
+      if(affectedRows) {
+        res.json({success: true, message: 'Member deleted'});
+      }
+      else {
+        var notFound = new Error('Member not found');
+        notFound.status = 404;
+        next(notFound);
+      }
+    })
+    .catch( (err) => {
+      next(new Error('Internal server error')); // return 500
+    });
 
 });
